fix: move quotes components under the components/quotes
