+++ examples/app-crm/src/components/index.ts
@@ -2,6 +2,5 @@ export * from "./project-kanban-card";
 export * from "./text";
 export * from "./fullscreen-loading";
 export * from "./deal-kanban-card";
-export * from "./quotes-form-modal";
-export * from "./quote-status-tag";
+export * from "./quotes";
 export * from "./custom-avatar";
