+++ examples/app-crm/src/components/quotes/index.ts
@@ -1,0 +1,6 @@
+export * from "./form-modal";
+export * from "./status-tag";
+export * from "./pdf-export";
+export * from "./products-services";
+export * from "./show-description";
+export * from "./status-indicator";
