+++ examples/app-crm/src/components/quotes/show-description.tsx
@@ -2,10 +2,10 @@ import { HttpError } from "@refinedev/core";
 import { useParams } from "react-router-dom";
 import { useForm } from "@refinedev/antd";
 import MDEditor from "@uiw/react-md-editor";
-import { Quote, QuoteUpdateInput } from "../../../interfaces/graphql";
+import { Quote, QuoteUpdateInput } from "../../interfaces/graphql";
 import { Form, Spin } from "antd";
 
-export const Description = () => {
+export const ShowDescription = () => {
     const params = useParams<{ id: string }>();
 
     const { formProps, queryResult, autoSaveProps } = useForm<
