+++ examples/app-crm/src/components/quotes/status-indicator/index.module.css
