+++ examples/app-crm/src/components/quotes/status-tag.tsx
